src/requirements.txt
beautifulsoup4
pandas
pdfreader
requests
tabula-py